requirements.txt
# SentinelAI — Python Dependencies

# Web Framework
fastapi==0.110.0
uvicorn[standard]==0.29.0

# Database
sqlalchemy==2.0.29
aiosqlite==0.20.0

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# ML
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.1
joblib==1.3.2

# HTTP (for geo API calls + scripts)
httpx==0.27.0
requests==2.31.0

# Email / OTP
secure-smtplib==0.1.1

# Environment
python-dotenv==1.0.1

# Utilities
pydantic==2.6.4
pydantic-settings==2.2.1
pytz==2024.1
